Solve second puzzle of third day
Change the solution in previous puzzle to instead use a make-shift set to hold
the coordinates of each point and its distance along the wire to the origin.
This solution is pretty bad, as the make-shift set is a list instead of the
binary tree in the implementation of Data.Set.

A better solution would have been to retrace the wire counting steps and check
each point against the set of intersection points. Then add the results for
each wire and obtain the minimum. Such a solution is left to other times. I
started to implement it but its getting late and I can't focus anymore. There's
a new puzzle tomorrow.