Solve first puzzle of tenth day
Read the input and parse it into a list of vectors centered on the top left
corner. Each vector represents the position of one asteroid in the input.
For each asteroid, change the referential of all asteroids to be centered on it
and normalize the vectors (using gcd instead of the norm).
Count all unobstructed asteroids which is now trivial, just count the number of
vectors with unique (x,y), disregarding the norm and get the maximum of this
process for all asteroids.