Solve second puzzle of tenth day
Not even I understand the mess this code is. Tired and overworked I managed
to compute and keep the angle of each vector in relation to (0, -1), the
vertical line.
Then, sort the vectors in ascending angle and go through this list, sending
any other vectors with the same angle to the back of the list to mimic the
obstruction of asteroids.