Solve first puzzle of eighteenth day
Query the intcode program on each point in a 50x50 square and sum the output,
as we want to know the total number of points affected by the beam and the
intcode output is 0 if the point is not affected, 1 if it is.

Also print the 50x50 square, which helps with the second part of the puzzle.